chatterbox_service/requirements.txt
chatterbox-tts>=0.1.7
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
